add new error + fix bugs
IncompleteMacroError now raised when the name of one argument in a macro is a command - meant to prevent unexpected errors (e.g. an argument cannot be called 'scoreboard')
Multiple towers caused an IndexError during macro substitution due to `substituted_lines` being defined outside of the loop and not being cleared - this is now fixed.